lang-chain/requirements.txt
langchain
langchain-core
langchain-community

langchain-google-genai

# LLM Providers

langchain-openai
langchain-mistralai
openai
mistralai

# Vector Database

chromadb

# Embedding Models

sentence-transformers

# Document Loaders dependencies

pypdf
unstructured
beautifulsoup4
lxml

# Text Processing

tiktoken

# Environment Variables

python-dotenv

# Data Handling

pandas
numpy

# API / App Layer (optional but common in RAG apps)

fastapi
uvicorn

# UI (optional)

streamlit

# Utilities

tqdm
requests

# Retrievers
arxiv

# Agents
duckduckgo-search 
langchain-experimental
ddgs
wikipedia

# langgraph-checkpoint-memory

fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-google-genai
langchain-mistralai
langchain-community
chromadb
